AI-Agent/requirements_free.txt
# FINTEL AI Agent - FREE Version Requirements

# LangChain with Ollama (FREE)
langchain==0.1.0
langchain-community==0.0.13

# OCR with Tesseract (FREE)
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.1.0
pdf2image==1.17.0

# Machine Learning (FREE)
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
scipy==1.11.4

# Database
pymongo==4.6.1

# API Server
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
requests==2.31.0
